//CRM-4543. //Block delete for activities which modify case attributes.
git-svn-id: http://svn.civicrm.org/civicrm/trunk@21593 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Case/Form/Activity.php
@@ -276,11 +276,22 @@ class CRM_Case_Form_Activity extends CRM_Activity_Form_Activity
     {
         if ( $this->_action & CRM_Core_Action::DELETE ) {
             $statusMsg = null;
-            $params = array( 'id' => $this->_activityId );
-            $activityDelete = CRM_Activity_BAO_Activity::deleteActivity( $params, true );
-            if ( $activityDelete ) {
-                $statusMsg = ts('The selected activity has been moved to the Trash. You can view and / or restore deleted activities by checking "Deleted Activities" from the Case Activities search filter (under Manage Case).<br />');
+          
+            //block deleting activities which affects
+            //case attributes.CRM-4543
+            $activityCondition = " AND v.name IN ('Open Case', 'Change Case Type', 'Change Case Status', 'Change Case Start Date')";
+            $caseAttributeActivities = CRM_Core_OptionGroup::values( 'activity_type', false, false, false, $activityCondition );
+            
+            if ( !array_key_exists($this->_activityTypeId, $caseAttributeActivities) ) {
+                $params = array( 'id' => $this->_activityId );
+                $activityDelete = CRM_Activity_BAO_Activity::deleteActivity( $params, true );
+                if ( $activityDelete ) {
+                    $statusMsg = ts('The selected activity has been moved to the Trash. You can view and / or restore deleted activities by checking "Deleted Activities" from the Case Activities search filter (under Manage Case).<br />');
+                }
+            } else {
+                $statusMsg = ts("Selected Activity cannot be deleted.");
             }
+            
             CRM_Core_Session::setStatus( $statusMsg );
             return;
         }
